auth_service/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
PyJWT>=2.8
bcrypt>=4.1
redis>=5.0
requests>=2.31
Pillow>=10.2
drf-spectacular>=0.27

gunicorn>=21.2.0

whitenoise>=6.6.0

whitenoise>=6.6.0
